+++ src/Identifiable.csproj
@@ -3,7 +3,7 @@
     <PropertyGroup>
         <LangVersion>latest</LangVersion>
         <TargetFramework>netstandard2.0</TargetFramework>
-        <Version>3.0.0-alpha01</Version>
+        <Version>3.0.0</Version>
         <PackageId>Identifiable</PackageId>
         <Authors>Sean Terry</Authors>
         <Description>Generates GUIDs based on current time (UUIDv1), MD5 hash (UUIDv3), and SHA1 hash (UUIDv5). Includes a non-standard variant layout of time-based GUIDs for near-sequential sorting in SQL Server.</Description>
@@ -12,7 +12,7 @@
         <Copyright>Copyright 2017 Sean Terry</Copyright>
         <PackageLicenseUrl>https://github.com/seanterry42/Identifiable/blob/master/LICENSE</PackageLicenseUrl>
         <PackageProjectUrl>https://github.com/seanterry42/Identifiable</PackageProjectUrl>
-        <DocumentationFile>bin\Release\netstandard2.0\Identifiable.xml</DocumentationFile>
+        <DocumentationFile>bin\$(Configuration)\netstandard2.0\Identifiable.xml</DocumentationFile>
     </PropertyGroup>
 
 </Project>
