--- src/Internal/TimeGuidFactory.cs
@@ -1,34 +1,0 @@
-﻿/*  Copyright 2017 Sean Terry
-
-    Licensed under the Apache License, Version 2.0 (the "License");
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-
-        http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
-*/
-
-using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Identifiable.Internal
-{
-    /// <summary>
-    /// Factory for generating time-based GUIDs.
-    /// </summary>
-    
-    public class TimeGuidFactory : ITimeGuidFactory
-    {
-        /// <summary>
-        /// Gets the start of the Gregorian calenadar, used per RFC-4122 for UUID Version 1.
-        /// </summary>
-
-        public static long Epoch { get; } = new DateTime( 1582, 10, 15 ).Ticks;
-    }
-}
